refactor: rethink app architecture
Relocate pure-domain helpers out of the UI tree into the domain layer,
enforcing the lib/-is-the-domain-layer boundary from CLAUDE.md.

- Move buildGenerationJob (builds a domain GenerationJob, imports only
  lib/) from app/components/canvas/utils/ to lib/generation/, next to the
  queue type it returns.
- Consolidate getElementCharacterNames into lib/canvas/characterNames.ts
  alongside its parseCharacterNames dependency, removing a cross-module hop.
- Move the corresponding tests into the matching lib/ __tests__ folders.

Only scrollToScene (a genuine DOM/UI helper) remains in the canvas utils
folder. No behavior change.
